requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
sqlalchemy==2.0.31
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.6.0
email-validator==2.2.0  